refactor getMeals function
