+++ src/index.js
@@ -1,7 +1,4 @@
 import './style.css';
 import getMeals from './getMeals';
-import mealList from './functions';
 
-window.addEventListener('load', async () => {
-  mealList(await getMeals());
-});
+getMeals();
